*** empty log message ***

+++ aubit4glsrc/trunk/incl/a4gl_libaubit4gl.h
@@ -24,7 +24,7 @@
 # | contact afalout@ihug.co.nz                                           |
 # +----------------------------------------------------------------------+
 #
-# $Id: a4gl_libaubit4gl.h,v 1.244 2006-09-13 15:07:16 briantan Exp $
+# $Id: a4gl_libaubit4gl.h,v 1.245 2006-09-14 13:26:23 briantan Exp $
 #
 */
 
@@ -1698,8 +1698,7 @@ void A4GL_set_ccnt(int a);
 
 
   /* ============================ pointers.c ============================= */
-//  int A4GL_has_pointer (char *pname, char t);
-  long A4GL_has_pointer (char *pname, char t);
+  int A4GL_has_pointer (char *pname, char t);
   void A4GL_add_pointer (char *orig_name, char type, void *ptr);
   void A4GL_del_pointer (char *pname, char t);
   int A4GL_find_pointer_ptr (char *name, char *type, void *ptr);
